NSAT-775 (#4)
* Commit dist folder and change main file

This will allow it to be pulled via npm directly from github with minimal
changes in the apps that use it.

* Prevent drag and drop if node is being renamed

This allows selected text with mouse curser when renaming a node

* Allow showing an icon to indicate item draggability.

* Add 'checked' class to allow custom styling of checked nodes.

* Rebuild dist

* Add prettier ignore dist folder

* Cherry pick pre-push hook from nabn:update-dist repo branch

* Fix pre-push hook on windows

"mkdir -p dist" is not valid on windows and creates
two folders -p/ and dist/

* Rebuild dist

* Update pre-push hook to commit dist folder.

* Add safeHtml filter for dragTemplate

* Don't fire node removed event when it was simply moved to another location

* Rebuild dist

* Ensure all files are included from the dist folder

* Fix moving a node

* Rebuild dist

* Fix issue where expading collapsed and checked item does not check it's children

Also fix a typo Childern -> Children

* Fix issues with checking/checking collapsed item

When item was collapsed, checked/unchecked status was not getting applied
to its children due to child controllers being deleted. This change will
prevent a controller being deleted if the item still exists in the parent.

* Rebuild dist

* Ensure dist folder changes are pushed

For some reason the commit created during pre-push does not get pushed.

* WIP: drag and drop multiple checked items

* Uncheck moved items

* Ensure moved items always have controller set.

Due to async loading of tree childred, in some cases when moving an item
its controller was getting deleted after being added, i.e. setController
was being called first on the item in the new tree, before the same item in
the old tree was destroyed and deleteController was called.

* Allow dragging unchecked node on its own leaving checked nodes as they are

* Rebuild dist

* Prevent pre-push loop

* NSAT-775 Only uncheck an item after moving it if it was checked.

This prevents false node unchecked events from being fired.

* NSAT-775 Allow dropping a node into a branch when both are siblings.

* NSAT-775 Move a node under dropped item.

* Rebuild dist

* Update repo urls to ArriaOpenSource

* NSAT-775 Allow moving node after sibling instead of swapping

This is configurable.
This commit also changes NodeMovedEvent to allow passing previous position
in parent, which is useful when nodes are swapped.

* Rebuild dist

* NSAT-775 Add docs for new moveSiblingAfter function

* NSAT-775 Allow different "drag-over" styling for branches and leaves.

* NSAT-775 WIP: allow dropping items under a branch (folder)

E.g. dropping a node between two bronches (folders).

* Rebuild dist

* NSAT-775 Allow selecting an item by clicking anywhere within value-container

* NSAT-775 allow dropping a node between two branches

instead of into the branch.

* Rebuild dist

* Revert "NSAT-775 allow dropping a node between two branches"

This reverts commit 8c35ca5cf5bfc21f2225b56f5c859caa9d6a59e3.

* Revert "NSAT-775 WIP: allow dropping items under a branch (folder)"

This reverts commit e644f6b7db49ce1582d4fef058d5a4e7f338d06b.

* Rebuild dist

* NSAT-775 Allow droping nodes above/below/into.

Leaf nodes now have two drop positions: above and below.
Branch nodes have three drop positions above, below and into.
This removes the option to swap nodes. Moving a node is now the only option.
Different styling can now also be applied for all three different drop positons,
as well as different styling for branch and leaf drop.

* Rebuild dist

* NSAT-775 Fix drop positioning when tree contains relative positioned elements

* Rebuild dist

* NSAT-775 Fix issue where node was being moved down two rows below instead of one.

* Rebuild dist

* Revert dc48fc1669ecc2c5f3085f3b023f894c976f6afc

Allow selecting an item by clicking anywhere within value-container

* Rebuild dist

* NSAT-775 Fix issue with moving items above/below

* NSAT-775 Prevent dropping a node on itself

* NSAT-775 Prevent moving a node for certain cases.

Prevent when:
* Moving into it's own parent
* Moving above/below/into itself.
* Moving below expanded branch (this can still be achieved by moving above the first item that is outside and immediately below the expanded branch).

* NSAT-775 Allow styling node that is being dragged

* NSAT-775 Fix issue where checked nodes were released prematurely

* Rebuild dist

* NSAT-775 WIP: Allow setting custom drag image.

* Rebuild dist

* NSAT-775 Prevent dropping selected items onto themselves

* NSAT-775 Uncheck parent when moving unchecked node

* NSAT-775 Move parent only when moving parent with children

* NSAT-775 Move checkboxes to the left-most

* Rebuild dist

* Expose draggable events

* Rebuild dist

* NSAT-775 Fix an issue where moving checked children would not move the parent.

* Fix typo

* Fix another type

* Fix the same typo in the rest of the code

* NSAT-775 Add ability to ignore children when parent is checked.

* NSAT-775 Fix issue with parent checked state not being correct

in some cases when moving nodes.

* NSAT-775 Add option to ignore parent when child is checked

i.e. don't check parent when all children are checked.

* Rebuild dist

* NSAT-775 Make indeterminate property public in NodeIndeterminateEvent

* Rebuild dist

* NSAT-775 Fix issues related to checking/unchecking items

* NSAT-775 Open menu upwards if close to screen bottom.

* NSAT-775 Open menu up or down based on available space below.

* Rebuild dist

* NSAT-775 Fix a case when there is not scroll parent

* Rebuild dist

* NSAT-775 Fix a case when we want to uncheck all children by unchecking a parent

* Rebuild dist

* NSAT-775 Fix e2e tests broken due to settings changes.

* NSAT-775 Fix broken tests

* Commit uncommited tests

* Remove undefined controler guard for when moving checked nodes.

* Rebuild dist

* Double click (#7)

* Add support for double-click event

Fixes #6

* Rebuild dist

+++ src/tree-internal.component.ts
@@ -21,9 +21,11 @@ import { NodeEditableEvent, NodeEditableEventAction } from './editable/editable.
 import { NodeCheckedEvent, NodeEvent } from './tree.events';
 import { TreeService } from './tree.service';
 import * as EventUtils from './utils/event.utils';
-import { NodeDraggableEvent } from './draggable/draggable.events';
+import { NodeDraggableEvent, DropPosition } from './draggable/draggable.events';
 import { Subscription } from 'rxjs/Subscription';
 import { get, isNil } from './utils/fn.utils';
+import { NodeDraggableService } from './draggable/node-draggable.service';
+import { CapturedNode } from './draggable/captured-node';
 
 @Component({
   selector: 'tree-internal',
@@ -31,21 +33,22 @@ import { get, isNil } from './utils/fn.utils';
   <ul class="tree" *ngIf="tree" [ngClass]="{rootless: isRootHidden()}">
     <li>
       <div class="value-container"
-        [ngClass]="{rootless: isRootHidden()}"
+        [ngClass]="{rootless: isRootHidden(), checked: tree.checked}"
         [class.selected]="isSelected"
         (contextmenu)="showRightMenu($event)"
         [nodeDraggable]="nodeElementRef"
         [tree]="tree">
 
-        <div class="folding" (click)="onSwitchFoldingType()" [ngClass]="tree.foldingCssClass"></div>
-
         <div class="node-checkbox" *ngIf="settings.showCheckboxes">
-        <input checkbox  type="checkbox" [disabled]="isReadOnly" [checked]="this.tree.checked" (change)="switchNodeCheckStatus()" #checkbox />
-         </div>
+          <input checkbox  type="checkbox" [disabled]="isReadOnly" [checked]="tree.checked" (change)="switchNodeCheckStatus()" #checkbox />
+        </div>
+
+        <div class="folding" (click)="onSwitchFoldingType()" [ngClass]="tree.foldingCssClass"></div>
 
         <div class="node-value"
           *ngIf="!shouldShowInputForTreeValue()"
           [class.node-selected]="isSelected"
+          (dblClick)="onNodeDoubleClicked()"
           (click)="onNodeSelected($event)">
             <div *ngIf="tree.nodeTemplate" class="node-template" [innerHTML]="tree.nodeTemplate | safeHtml"></div>
             <span *ngIf="!template" class="node-name" [innerHTML]="tree.value | safeHtml"></span>
@@ -63,6 +66,7 @@ import { get, isNil } from './utils/fn.utils';
         <node-menu *ngIf="tree.hasLeftMenu() && isLeftMenuVisible && !hasCustomMenu()"
           (menuItemSelected)="onMenuItemSelected($event)">
         </node-menu>
+        <div class="drag-template" *ngIf="tree.hasDragIcon()" [innerHTML]="tree.dragTemplate | safeHtml"></div>
       </div>
 
       <node-menu *ngIf="isRightMenuVisible && !hasCustomMenu()"
@@ -104,6 +108,7 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
   public constructor(
     private nodeMenuService: NodeMenuService,
     public treeService: TreeService,
+    public nodeDraggableService: NodeDraggableService,
     public nodeElementRef: ElementRef
   ) {}
 
@@ -139,12 +144,28 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
 
     this.subscriptions.push(
       this.treeService.draggedStream(this.tree, this.nodeElementRef).subscribe((e: NodeDraggableEvent) => {
-        if (this.tree.hasSibling(e.captured.tree)) {
-          this.swapWithSibling(e.captured.tree, this.tree);
-        } else if (this.tree.isBranch()) {
-          this.moveNodeToThisTreeAndRemoveFromPreviousOne(e, this.tree);
-        } else {
-          this.moveNodeToParentTreeAndRemoveFromPreviousOne(e, this.tree);
+        // Remove child nodes if parent is being moved (child nodes will move with the parent)
+        const nodesToMove = e.captured.filter(cn => !cn.tree.parent.checked);
+
+        let i = nodesToMove.length;
+        while (i--) {
+          const node = nodesToMove[i];
+          const ctrl = this.treeService.getController(node.tree.id);
+          if (ctrl.isChecked()) {
+            ctrl.uncheck();
+          }
+
+          if (this.tree.isBranch() && e.position === DropPosition.Into) {
+            this.moveNodeToThisTreeAndRemoveFromPreviousOne(node.tree, this.tree);
+          } else if (this.tree.hasSibling(node.tree)) {
+            this.moveSibling(node.tree, this.tree, e.position);
+          } else {
+            this.moveNodeToParentTreeAndRemoveFromPreviousOne(node.tree, this.tree, e.position);
+          }
+        }
+        const parentCtrl = this.treeService.getController(this.tree.parent.id);
+        if (parentCtrl) {
+          parentCtrl.updateCheckboxState();
         }
       })
     );
@@ -162,28 +183,53 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
   }
 
   public ngOnDestroy(): void {
-    if (get(this.tree, 'node.id', '')) {
+    if (get(this.tree, 'node.id', '') && !(this.tree.parent && this.tree.parent.children.indexOf(this.tree) > -1)) {
       this.treeService.deleteController(this.tree.node.id);
     }
 
     this.subscriptions.forEach(sub => sub && sub.unsubscribe());
   }
 
-  private swapWithSibling(sibling: Tree, tree: Tree): void {
-    tree.swapWithSibling(sibling);
-    this.treeService.fireNodeMoved(sibling, sibling.parent);
+  private moveSibling(sibling: Tree, tree: Tree, position: DropPosition): void {
+    const previousPositionInParent = sibling.positionInParent;
+    if (position === DropPosition.Above) {
+      tree.moveSiblingAbove(sibling);
+    } else {
+      tree.moveSiblingBelow(sibling);
+    }
+    this.treeService.fireNodeMoved(sibling, sibling.parent, previousPositionInParent);
   }
 
-  private moveNodeToThisTreeAndRemoveFromPreviousOne(e: NodeDraggableEvent, tree: Tree): void {
-    this.treeService.fireNodeRemoved(e.captured.tree);
-    const addedChild = tree.addChild(e.captured.tree);
-    this.treeService.fireNodeMoved(addedChild, e.captured.tree.parent);
+  private moveNodeToThisTreeAndRemoveFromPreviousOne(capturedTree: Tree, moveToTree: Tree): void {
+    capturedTree.removeItselfFromParent();
+    setTimeout(() => {
+      const addedChild = moveToTree.addChild(capturedTree);
+      this.treeService.fireNodeMoved(addedChild, capturedTree.parent);
+    });
   }
 
-  private moveNodeToParentTreeAndRemoveFromPreviousOne(e: NodeDraggableEvent, tree: Tree): void {
-    this.treeService.fireNodeRemoved(e.captured.tree);
-    const addedSibling = tree.addSibling(e.captured.tree, tree.positionInParent);
-    this.treeService.fireNodeMoved(addedSibling, e.captured.tree.parent);
+  private moveNodeToParentTreeAndRemoveFromPreviousOne(
+    capturedTree: Tree,
+    moveToTree: Tree,
+    position: DropPosition
+  ): void {
+    capturedTree.removeItselfFromParent();
+    setTimeout(() => {
+      let insertAtIndex = moveToTree.positionInParent;
+      if (position === DropPosition.Below) {
+        insertAtIndex++;
+      }
+      const addedSibling = moveToTree.addSibling(capturedTree, insertAtIndex);
+      this.treeService.fireNodeMoved(addedSibling, capturedTree.parent);
+    });
+  }
+
+  public onNodeDoubleClicked(): void {
+    if (!this.tree.selectionAllowed) {
+      return;
+    }
+
+    this.treeService.fireNodeDoubleClicked(this.tree);
   }
 
   public onNodeSelected(e: { button: number }): void {
@@ -323,30 +369,51 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
     }
   }
 
-  public onNodeChecked(): void {
+  public onNodeChecked(ignoreChildren: boolean = false): void {
     if (!this.checkboxElementRef) {
       return;
     }
 
-    this.checkboxElementRef.nativeElement.indeterminate = false;
-    this.treeService.fireNodeChecked(this.tree);
-    this.executeOnChildController(controller => controller.check());
-    this.tree.checked = true;
+    if (!this.tree.checked) {
+      this.nodeDraggableService.addCheckedNode(new CapturedNode(this.nodeElementRef, this.tree));
+      this.onNodeIndeterminate(false);
+      this.tree.checked = true;
+      this.treeService.fireNodeChecked(this.tree);
+    }
+
+    if (!ignoreChildren) {
+      this.executeOnChildController(controller => controller.check());
+    }
   }
 
-  public onNodeUnchecked(): void {
+  public onNodeUnchecked(ignoreChildren: boolean = false): void {
     if (!this.checkboxElementRef) {
       return;
     }
 
-    this.checkboxElementRef.nativeElement.indeterminate = false;
-    this.treeService.fireNodeUnchecked(this.tree);
-    this.executeOnChildController(controller => controller.uncheck());
-    this.tree.checked = false;
+    if (this.tree.checked) {
+      this.nodeDraggableService.removeCheckedNodeById(this.tree.id);
+      this.onNodeIndeterminate(false);
+      this.tree.checked = false;
+      this.treeService.fireNodeUnchecked(this.tree);
+    }
+
+    if (!ignoreChildren) {
+      this.executeOnChildController(controller => controller.uncheck());
+    }
+  }
+
+  public onNodeIndeterminate(indeterminate: boolean): void {
+    if (!this.checkboxElementRef || this.checkboxElementRef.nativeElement.indeterminate === indeterminate) {
+      return;
+    }
+
+    this.checkboxElementRef.nativeElement.indeterminate = indeterminate;
+    this.treeService.fireNodeIndeterminate(this.tree, indeterminate);
   }
 
   private executeOnChildController(executor: (controller: TreeController) => void) {
-    if (this.tree.hasLoadedChildern()) {
+    if (this.tree.hasLoadedChildren()) {
       this.tree.children.forEach((child: Tree) => {
         const controller = this.treeService.getController(child.id);
         if (!isNil(controller)) {
@@ -361,17 +428,18 @@ export class TreeInternalComponent implements OnInit, OnChanges, OnDestroy, Afte
     setTimeout(() => {
       const checkedChildrenAmount = this.tree.checkedChildrenAmount();
       if (checkedChildrenAmount === 0) {
-        this.checkboxElementRef.nativeElement.indeterminate = false;
-        this.tree.checked = false;
-        this.treeService.fireNodeUnchecked(this.tree);
+        this.onNodeUnchecked(true);
+        this.onNodeIndeterminate(false);
       } else if (checkedChildrenAmount === this.tree.loadedChildrenAmount()) {
-        this.checkboxElementRef.nativeElement.indeterminate = false;
-        this.tree.checked = true;
-        this.treeService.fireNodeChecked(this.tree);
+        if (!this.settings.ignoreParentOnCheck) {
+          this.onNodeChecked(true);
+          this.onNodeIndeterminate(false);
+        } else if (!this.tree.checked) {
+          this.onNodeIndeterminate(true);
+        }
       } else {
-        this.tree.checked = false;
-        this.checkboxElementRef.nativeElement.indeterminate = true;
-        this.treeService.fireNodeIndetermined(this.tree);
+        this.onNodeUnchecked(true);
+        this.onNodeIndeterminate(true);
       }
     });
   }
